+++ semanas-20e21/FutureTube-Front/src/containers/SignUpPage/index.js
@@ -1,28 +1,12 @@
-// TODO: Isolar o form de cadastro em um componente separado
 // TODO: Adicionar url da phto no serviço de cadastro de usuário
-
 
 import React, { useState } from "react";
 import { connect } from 'react-redux';
 import { push } from 'connected-react-router';
 import { routes } from '../Router';
-import MainButtonComponent from '../../components/MainButton';
-import Logo from "../../components/Logo";
-import { 
-    SignUpWrapper,
-    ContainerSignUpPage, 
-    TextRegister, 
-    InputName, 
-    InputEmail,
-    InputBirth, 
-    InputPassword, 
-    InputPasswordConfirmation 
-} from './styled';
-import Visibility from '@material-ui/icons/Visibility';
-import VisibilityOff from '@material-ui/icons/VisibilityOff';
-import InputAdornment from '@material-ui/core/InputAdornment';
-import IconButton from '@material-ui/core/IconButton';
+import { SignUpWrapper } from './styled';
 import * as firebase from 'firebase';
+import SignUpForm from '../../components/SignUpForm';
 
 const SignUpPage = props => {
     const [userName, setUserName] = useState('');
@@ -32,10 +16,7 @@ const SignUpPage = props => {
     const [passwordConfirmation, setPasswordConfirmation] = useState('');
     const [showPassword, setShowPassword] = useState(false);
     const [showPasswordConfirmation, setShowPasswordConfirmation] = useState(false);
-    const [passwordErrorMessage, setPasswordErrorMessage] = useState({
-        msg: '', 
-        status: false
-    });
+    const [passwordErrorMessage, setPasswordErrorMessage] = useState({ msg: '', status: false });
     const [heightInputPasswordConfirmation, setHeightInputPasswordConfirmation] = useState('56px');
 
     const handleFieldChangeUserName = event => {
@@ -109,117 +90,29 @@ const SignUpPage = props => {
 
     return ( 
         <SignUpWrapper>
-            <ContainerSignUpPage onSubmit={handleSubmit}>
-                <Logo />
-                <TextRegister>Cadastrar</TextRegister>
-                <InputName
-                    required
-                    id="outlined-required-name"
-                    label="Nome"
-                    placeholder="Nome e Sobrenome"
-                    margin="normal"
-                    variant="outlined"
-                    InputLabelProps={{
-                        shrink: true,
-                    }}
-                    onChange={handleFieldChangeUserName}
-                    name="name"
-                    type="text"
-                    value={userName}
-                />
-                <InputEmail
-                    required
-                    id="outlined-required-email"
-                    label="E-mail"
-                    placeholder="email@email.com"
-                    margin="normal"
-                    variant="outlined"
-                    InputLabelProps={{
-                        shrink: true,
-                    }}
-                    onChange={handleFieldChangeEmail}
-                    name="email"
-                    type="email"
-                    value={email}
-                />
-                <InputBirth
-                    required
-                    id="date"
-                    label="Birthday"
-                    type="date"
-                    InputLabelProps={{
-                    shrink: true,
-                    }}
-                    onChange={handleFieldChangeBirth}
-                    value={birth}
-                />
-                <InputPassword
-                    required
-                    id="outlined-required-password-signUp"
-                    label="Senha"
-                    placeholder="Mínimo 6 caracteres"
-                    margin="normal"
-                    variant="outlined"
-                    InputLabelProps={{
-                        shrink: true,
-                    }}
-                    InputProps={{
-                        endAdornment: (
-                            <InputAdornment position="end">
-                                <IconButton
-                                    aria-label="Toggle password visibility"
-                                    onClick={handleClickShowPassword}
-                                >
-                                    {showPassword ? <VisibilityOff /> : <Visibility />}
-                                </IconButton>
-                            </InputAdornment>
-                        ),
-                    }}
-                    inputProps={{
-                        minlength: "6"
-                    }}
-                    onChange={handleFieldChangePassword}
-                    name="passwordSignUp"
-                    type={showPassword ? 'text' : 'password'}
-                    value={password}
-                    title="A senha deve ter no mínimo 6 caracteres"
-                />
-                <InputPasswordConfirmation
-                    required
-                    id="outlined-required-confirm"
-                    label="Confirmar"
-                    placeholder="Confirme a senha anterior"
-                    margin="normal"
-                    variant="outlined"
-                    height={heightInputPasswordConfirmation}
-                    error={passwordErrorMessage.status}
-                    helperText={passwordErrorMessage.msg}
-                    InputLabelProps={{
-                        shrink: true,
-                    }}
-                    InputProps={{
-                        endAdornment: (
-                            <InputAdornment position="end">
-                                <IconButton
-                                    aria-label="Toggle password visibility"
-                                    onClick={handleClickShowPasswordConfirmation}
-                                >
-                                    {showPasswordConfirmation ? <VisibilityOff /> : <Visibility />}
-                                </IconButton>
-                            </InputAdornment>
-                        ),
-                    }}
-                    onChange={handleFieldChangePasswordConfirmation}
-                    name="passwordConfirmation"
-                    type={showPasswordConfirmation ? 'text' : 'password'}
-                    value={passwordConfirmation}
-                />
-                <MainButtonComponent title="Criar" type="submit" />
-            </ContainerSignUpPage>
+            <SignUpForm
+                handleFieldChangeUserName={handleFieldChangeUserName}
+                handleFieldChangeEmail={handleFieldChangeEmail}
+                handleFieldChangeBirth={handleFieldChangeBirth}
+                handleFieldChangePassword={handleFieldChangePassword}
+                handleFieldChangePasswordConfirmation={handleFieldChangePasswordConfirmation}
+                handleClickShowPassword={handleClickShowPassword}
+                handleClickShowPasswordConfirmation={handleClickShowPasswordConfirmation}
+                handleSubmit={handleSubmit}
+                userName={userName}
+                email={email}
+                birth={birth}
+                showPassword={showPassword}
+                password={password}
+                height={heightInputPasswordConfirmation}
+                errorMessage={passwordErrorMessage.status}
+                passwordErrorMessage={passwordErrorMessage.msg}
+                showPasswordConfirmation={showPasswordConfirmation}
+                passwordConfirmation={passwordConfirmation}
+            />
         </SignUpWrapper>
     );
 };
-
 
 const mapDispatchToProps = (dispatch) => {
     return {
